requirements.txt
click>=6.7,<7.0
attrs>=18.0,<19.0
